Created initial spring boot folder structure (#15)
