+++ fullstack/generator/springboot/template/content_structure/model/__entity_name__.java.jinja
@@ -2,9 +2,9 @@ package {{ group_name }}.{{ app_name | lower }}.model;
 
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+import lombok.Data;
 
-import javax.persistence.*;
-import java.util.Set;
+import jakarta.persistence.*;
 
 @Entity
 @Data
