+++ fullstack/generator/springboot/template/main_file/__app_name_cap__Application.java.jinja
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class {{ app_name | capitalize }}Application {
+public class {{ app_name_cap }}Application {
 
 	public static void main(String[] args) {
-		SpringApplication.run({{ app_name | capitalize }}Application.class, args);
+		SpringApplication.run({{ app_name_cap }}Application.class, args);
 	}
 
 }
