+++ fullstack/generator/springboot/template/springboot_structure/__app_name_lower__/src/main/resources/application.properties.jinja
@@ -1,0 +1,2 @@
+spring.application.name={{ app_name }}
+server.port=8442
