+++ fullstack/generator/springboot/template/springboot_structure/__app_name_lower__/src/test/resources/application-test.properties.jinja
@@ -1,0 +1,1 @@
+spring.application.name={{ app_name }}
